+++ CHANGELOG.md
@@ -3,6 +3,7 @@ AppEngine version, listed here to ease deployment and troubleshooting.
 
 ## Next Release (replace with git tag when deployed)
  * NOTE: renamed `is:dart3-ready` to `is:dart3-compatible`.
+ * Note: started to delete leftover objects in the public bucket.
 
 ## `20230307t120800-all`
  * Bumped runtimeVersion to `2023.03.07`.